rename main dashboard file name, move component modal create feature plan to page
